BUG: fix the rendering of 2D gaussian distributions
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=10054
git-svn-id: http://svn.slicer.org/Slicer4/trunk@10054 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/Wizard/vtkEMSegmentIntensityNormalizationStep.cxx
@@ -1100,7 +1100,7 @@ void vtkEMSegmentIntensityNormalizationStep::ProcessCursorMovingGUIEvents(
     this->NormalizationHistogram->GetRange(range);
     occurence = this->NormalizationHistogram->GetOccurenceAtValue(90);
 
-    for (int i = this->NormalizationPiecewiseFunctionEditor->
+    for (double i = this->NormalizationPiecewiseFunctionEditor->
         GetParameterCursorPosition(); i < range[1] ; i++)
     {
       double num = this->NormalizationHistogram->GetOccurenceAtValue(i);
